requirements.txt
joblib==1.5.2
matplotlib==3.10.6
numpy==2.3.3
pandas==2.3.2
Pillow==11.3.0
streamlit==1.40.0
torch==2.8.0
torchvision==0.23.0
